Added a shotgun to the playercontroller
Implemented a shotgun shooting into the player controller. It is nested in the shoot enumerator

+++ Assets/Scripts/IDamage.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public interface IDamage 
 {
-    void takeDamage(int dmg);
+    void takeDamage(float dmg);
 
 }
